backend/qualla/requirements.txt
asgiref==3.5.2
Django==4.1.2
pytz==2022.5
sqlparse==0.4.3
djangorestframework==3.11.1
coverage==6.5
pylint==2.15.5
pylint-django==2.5.3
